Removed 'Partnerzy' tab and added content to 'O Poli'
